Lots of work done.  Now Firebird, Thunderbird, and Mozilla are all supported by TagZilla.  A lot of work has gone into the prefs mechanism; it's all one window now, and it's the same window, no matter which browser you use. Huge props go out to optimoz (MozGest) and its coders for their method (and code!) of doing the prefs cross-platform like that.  I love it.
Another big change is in the installer.  I borrowed the generic install.js
used in the offline extension provided by mozilla.org, so it can now install
TagZilla globally or to your personal directory only.  The downside to this
change is that TagZilla is now a jar-based, not directory-based, installation.
Sorry.  (The installer has been tested in Firebird and Thunderbird, but NOT
in Mozilla yet!)

Will release soon, but not quite yet.  Some issues I want to take care of
first:
- the toolbar icons in Firebird and Thunderbird don't always fit right
- said icons are fugly in the toolbar palette, before you put them on a toolbar
- it's about time I took care of the dos/unix/mac newline problem

Keep well folks!  Thanks a ton for your patience.