Bunch of work on some hopeful bugfixes.  The most obvious is that, if you set it to auto-insert a tagline without popping up the window, the tagline gets added to the message when the compose window opens, so you can look at it and stuff, rather than the old behaviour where it added the tagline when you pressed Send.  (If you don't have the TagZilla window turned off for this, it still appears when you press Send.)
Also added is experimental UTF-8 textfile support.  Thanks to Rob Ginda
(and all his helper elves) for writing ChatZilla, which contained useful
functions that I was able to use for this purpose!  See bug 2098.

Finally, check it for the crashing bug.  I didn't do anything about it
directly, but it's not happening to me any more.  Maybe it was just bit
rot...